Squashed commit of the following:
commit bb8c6bae8317f20ebfdb0b8d9aca603806725cbe
Author: gc <30398469+gc@users.noreply.github.com>
Date:   Wed Jun 23 05:23:13 2021 +1000

    give top collector to rank 1 overall cl

commit a14951d8a017bf12c800b6377cabd54a990ec232
Author: gc <30398469+gc@users.noreply.github.com>
Date:   Wed Jun 23 05:18:21 2021 +1000

    dont use warrior guild tokens if has attack cape

commit ebf9eb9c5f9e5740a5d24c876bb6f824d15d1035
Author: GC <30398469+gc@users.noreply.github.com>
Date:   Wed Jun 23 05:08:35 2021 +1000

    Fix similar items (#2168)

    * ?

    * fix dat

    * add more tests

    * add trailblazer graceful

commit 0f437f40e0e3474433824cd92852df0b2c55c12e
Author: gc <30398469+gc@users.noreply.github.com>
Date:   Wed Jun 23 03:47:29 2021 +1000

    refactor to dev instant trips

commit 93bc92ffe79db357ce5ab23a0bb45e7fced45b6a
Author: gc <30398469+gc@users.noreply.github.com>
Date:   Wed Jun 23 03:44:16 2021 +1000

    fix typo in quest cmd

commit 092ef6bca0a3b25774cdc0f7c83d2229e97b8367
Author: gc <30398469+gc@users.noreply.github.com>
Date:   Wed Jun 23 03:44:07 2021 +1000

    fix lb pets

commit c4b355a9904622c8cc28c451b5101ce9e699a10b
Author: gc <30398469+gc@users.noreply.github.com>
Date:   Wed Jun 23 03:44:01 2021 +1000

    dont give normal raids kc for cm

commit a37b6c1ddec50e052881953ce09820cc4c411a0a
Author: gc <30398469+gc@users.noreply.github.com>
Date:   Wed Jun 23 03:43:53 2021 +1000

    allow to equip cm pets

commit 2658ab25e25cbdfa16937a39a93c04825fb372a2
Author: Lily <Lilylicious@users.noreply.github.com>
Date:   Tue Jun 22 19:38:00 2021 +0200

    Refactor craft cmd, fix filters (#2150)

    * Craft bank usage and fletching filter

    * Round 1 review

    * Minor fix

    * Better GP handling and linting

    * Removed unreachable code

    * Removed missed GP code


commit 0462e743e795c0c95731268140b994e646b70eba
Author: gc <30398469+gc@users.noreply.github.com>
Date:   Wed Jun 23 02:30:46 2021 +1000

    add github issues search cmd

commit 0b596997ff6ec09733aa196dd65998c49a96d9b7
Author: GC <30398469+gc@users.noreply.github.com>
Date:   Wed Jun 23 02:19:43 2021 +1000

    Fix autoequip bypassing stat reqs (#2164)

commit e31defc86033b19a51f28fef438747799b37f356
Author: gc <30398469+gc@users.noreply.github.com>
Date:   Wed Jun 23 01:04:49 2021 +1000

    save activity after marking users as busy to avoid race condition

commit 0fe0494509581c31347c7ffbe69946967102a7a1
Author: Lily <Lilylicious@users.noreply.github.com>
Date:   Tue Jun 22 01:11:13 2021 +0200

    Give higher jad pet droprate on task (#2158)

commit 91aa431002b1dd9836429295907f4c70d7ba422f
Author: 6enny6 <76587427+6enny6@users.noreply.github.com>
Date:   Tue Jun 22 00:07:05 2021 +0100

    Diary fixes, add diary filter (#2139)

    * Update diaries.ts

    * Update diaries.ts

    * Update filterables.ts

    * Update filterables.ts


commit edb26206835a674afc249d44d65f5efe94906f8d
Author: themrrobert <10122432+themrrobert@users.noreply.github.com>
Date:   Mon Jun 21 16:04:38 2021 -0700

    Fix Black Mask CL and enhance the loot filter (#2142)

    * Fix black mask + filter loot cl

    * Minor formatting updates

    * Balancing abby specs


commit 13ad4ac524b5b92df4fec3dd3b953a08d3d5abaf
Author: gc <30398469+gc@users.noreply.github.com>
Date:   Tue Jun 22 08:26:54 2021 +1000

    remove all usages of hasItemEquipped and hasGearEquipped

commit 7a656433f2c829ee6feee7edb15acd4705663814
Author: gc <30398469+gc@users.noreply.github.com>
Date:   Tue Jun 22 08:17:02 2021 +1000

    remove all usages of hasArrayOfItemsEquipped

commit 20bebd35f03457e97bbd29d080e2eb6ce8674485
Author: gc <30398469+gc@users.noreply.github.com>
Date:   Tue Jun 22 08:11:27 2021 +1000

    fix continuing dark altar trips

commit e518eaae59c1da4b5fe32bb11a72e3817f415240
Author: gc <30398469+gc@users.noreply.github.com>
Date:   Tue Jun 22 06:58:01 2021 +1000

    fix unequip bug

commit 5c37cba0d6c234c37b33b31fa55e10c136da6ed6
Author: GC <30398469+gc@users.noreply.github.com>
Date:   Tue Jun 22 02:47:11 2021 +1000

    Gear improvements (#2153)

    * add tests for similar items gear equip check

    * add another test

    * add hasequipped rp cmd

commit 483bf131bfb55b9fc05d8640e9c61bd463fd6ac5
Author: Lily <Lilylicious@users.noreply.github.com>
Date:   Mon Jun 21 16:51:39 2021 +0200

    Fix not enough item error in craft cmd (#2147)

+++ src/commands/Minion/warriorsguild.ts
@@ -110,15 +110,17 @@ export default class extends BotCommand {
 		}
 
 		if (minigame === 'cyclops') {
+			const hasAttackCape = msg.author.getGear('melee').hasEquipped('Attack cape');
 			const maxTripLength = msg.author.maxTripLength(Activity.Cyclops);
 			// Check if either 100 warrior guild tokens or attack cape (similar items in future)
 			const amountTokens = userBank.amount('Warrior guild token');
-			if (amountTokens < 100) {
+			if (!hasAttackCape && amountTokens < 100) {
 				return msg.send('You need atleast 100 Warriors guild tokens to kill Cyclops.');
 			}
 			// If no quantity provided, set it to the max.
 			if (quantity === null) {
-				const maxTokensTripLength = Math.floor((amountTokens - 10) / 10) * Time.Minute;
+				const maxTokensTripLength =
+					Math.floor((hasAttackCape ? 100_000 : amountTokens - 10) / 10) * Time.Minute;
 				quantity = Math.floor(
 					(maxTokensTripLength > maxTripLength ? maxTripLength : maxTokensTripLength) / (Time.Second * 30)
 				);
@@ -138,7 +140,7 @@ export default class extends BotCommand {
 
 			const tokensToSpend = Math.floor((duration / Time.Minute) * 10 + 10);
 
-			if (amountTokens < tokensToSpend) {
+			if (!hasAttackCape && amountTokens < tokensToSpend) {
 				return msg.send(
 					`You don't have enough Warrior guild tokens to kill cyclopes for ${formatDuration(
 						duration
@@ -158,11 +160,15 @@ export default class extends BotCommand {
 
 			let response = `${
 				msg.author.minionName
-			} is now off to kill ${quantity}x Cyclops, it'll take around ${formatDuration(
-				duration
-			)} to finish. Removed ${tokensToSpend} Warrior guild tokens from your bank.`;
+			} is now off to kill ${quantity}x Cyclops, it'll take around ${formatDuration(duration)} to finish. ${
+				hasAttackCape
+					? 'You used no warrior guild tokens because you have an Attack cape.'
+					: `Removed ${tokensToSpend} Warrior guild tokens from your bank.`
+			}`;
 
-			await msg.author.removeItemFromBank(itemID('Warrior guild token'), tokensToSpend);
+			if (!hasAttackCape) {
+				await msg.author.removeItemFromBank(itemID('Warrior guild token'), tokensToSpend);
+			}
 
 			return msg.send(response);
 		}
